pyproject.toml
[tool.poetry]
name = "audio-analyzer"
version = "0.1.0"
description = ""
authors = ["imtrisha <imtrisha@randomdevs.org>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12.7"
python-ffmpeg = "^2.0.12"
speechrecognition = {extras = ["whisper-api"], version = "^3.11.0"}
google-cloud-speech = "^2.28.0"
vosk = "^0.3.45"
pocketsphinx = "^5.0.3"
soundfile = "^0.12.1"
torch = "^2.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
